documentation_backend/requirements.txt
alabaster==1.0.0
amqp==5.3.1
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.0
attrs==25.3.0
autolab-core==1.1.1
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.8.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
comm==0.1.4
contourpy==1.1.1
cryptography==45.0.3
cssselect2==0.8.0
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
Django==5.1.8
django-allauth==65.6.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
docutils==0.21.2
drf-yasg==1.21.10
embreex==2.17.7.post6
exceptiongroup==1.3.0
executing==2.0.0
filemagic==1.6
fonttools==4.58.2
freetype-py==2.5.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imagesize==1.4.1
inflection==0.5.1
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
kombu==5.5.3
lazy_loader==0.4
lxml==5.4.0
manifold3d==3.0.1
mapbox_earcut==1.0.3
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
multiprocess==0.70.17
nest-asyncio==1.5.8
networkx==3.4.2
numpy==1.26.0
opencv-python==4.11.0.86
oscrypto==1.3.0
packaging==23.2
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydyf==0.11.0
pyglet==2.1.3
Pygments==2.19.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyOpenGL==3.1.0
pyparsing==3.1.1
pypdf==5.5.0
PyPDF2==3.0.1
pyphen==0.17.2
pyreadline==2.1
pyrender==0.1.45
python-bidi==0.6.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.36.2
reportlab==4.4.1
requests==2.31.0
rpds-py==0.24.0
rtree==1.4.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setproctitle==1.3.5
shapely==2.0.7
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.1.3
sphinx-autodoc-typehints==3.0.1
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stack-data==0.6.3
svg.path==6.3
svglib==1.5.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.11.2
trimesh==4.6.6
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==2.4.0
vhacdx==0.0.8.post2
vine==5.1.0
visualization==1.0.0
wcwidth==0.2.8
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
xxhash==3.5.0
zopfli==0.2.3.post1
